==> code/2 - data_wrangling.Rmd <==
---
title: "2 - data_wrangling"
author: "Luis Segura"
date: "`r Sys.Date()`"
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)

options(tinytex.verbose = TRUE, scipen = 999)

mypckgs <- c("tidyverse", "here")

for (package in mypckgs) {
  library(package, character.only = T)
}

load(here("data/", "ensanut18_all.RData"))
```

## Limpieza y recodificacion de variables

```{r}
ensanut18_4analysis <- ensanut18_all |>
  labelled::remove_labels()  |>
  mutate(ses = factor(case_when(nsef == 1 ~ 3, 
                                nsef == 2 ~ 2, 
                                nsef == 3 ~ 1), 
                      labels = c("High", "Middle", "Low")), 
         age_cat = factor(ifelse(edad < 40, 1, 
                                 ifelse(edad > 39 & edad < 60, 2, 3)), 
                          labels = c("20 - 39 yo", "40 - 59 yo", "60+ yo")), 
         sexo = factor(sexo, labels = c("Men", "Women")), 
         urbanicity = factor(dominio, labels = c("Urban", "Rural")),  
         region = factor(region, labels = c("North", "Center", "Mexico City", "South")),
         region = fct_relevel(region, "Center"),
         l_indigena = factor(ifelse(l_indigena == 2, 0, l_indigena), labels = c("No", "Yes")), 
         usa_lentes = factor(ifelse(usa_lentes == 2, 0, usa_lentes), labels = c("No", "Yes")),
         disc_visual = factor(disc_visual, labels = c("No difficulty", "Some difficulty", 
                                                      "A lot of difficulty", "Cannot see at all")),
         disc_visual_severa = factor(ifelse(disc_visual %in% c("No difficulty", "Some difficulty"), 0, 1), labels = c("No", "Yes")),
         disc_visual_severa_num = ifelse(disc_visual %in% c("No difficulty", "Some difficulty"), 0, 1),
         p2_1_1 = p2_1_1 - 1, 
         p2_1_2 = p2_1_2 - 1, 
         p2_1_3 = p2_1_3 - 1, 
         p2_1_4 = p2_1_4 - 1, 
         p2_1_5 = p2_1_5 - 1, 
         p2_1_6_r = case_when(p2_1_6 == 4 ~ 3, 
                              p2_1_6 == 3 ~ 2, 
                              p2_1_6 == 2 ~ 1, 
                              p2_1_6 == 1 ~ 0), 
         p2_1_7 = p2_1_7 - 1, 
         cesd_score = rowSums(pick(p2_1_1, p2_1_2, p2_1_3, p2_1_4, p2_1_5, p2_1_6_r, p2_1_7)), 
         cesd = factor(ifelse((cesd_score > 8 & edad < 60) | (cesd_score > 4 & edad > 59), 1, 0), 
                       labels = c("No depressive symptoms", "Depressive symptoms")),
          
          
         insurance = factor(ifelse(p3_10_99 == 1, NA, 
                                   ifelse(p3_10_11 == 1, 0, 1)), labels = c("No insurance", "Insurance")), 
         work_pwk = factor(case_when(trabajo == 2 ~ 1, 
                                     trabajo == 1 ~ 0), labels = c("Yes", "No")),
         nivel = as.integer(nivel),
         education = factor(case_when(nivel < 3 ~ 4, 
                                      nivel == 3 ~ 3, 
                                      nivel == 4 ~ 2, 
                                      nivel > 4 ~ 1), 
                            labels = c("More than high school education", 
                                       "High school education","Secondary school", 
                                       "Elementary education or lower")), 
         suicide_int = factor(ifelse(p12_8 == 3, 0, 1), labels = c("No", "Si")), 
         suicide_int_py = factor(ifelse(p12_8_1 == 1, 1, 
                                 ifelse(p12_8_1 == 2, 0, NA)), labels = c("No", "Si")), 
         suicide_int_py_r = factor(case_when(suicide_int == "No" ~ 0, 
                                             suicide_int == "Si" & p12_8_1 == 2 ~ 0, 
                                             p12_8_1 == 1 ~ 1, 
                                             suicide_int == "Si" & is.na(p12_8_1) ~ NA), labels = c("No", "Si")), 
         cesd_num = ifelse(cesd == "Depressive symptoms", 1, 
                           ifelse(cesd == "No depressive symptoms", 0, NA)), 
         suicide_int_py_r_num = ifelse(suicide_int_py_r == "No", 0, 
                                     ifelse(suicide_int_py_r == "Si", 1, NA)), 
         suicide_int_num = ifelse(suicide_int == "No", 0, 1), 
         suicide_ideation_num = ifelse(p12_7 == 8, NA, 
                                   ifelse(p12_7 == 2, 0, p12_7)), 
         suicide_ideation = factor(suicide_ideation_num, labels = c("No", "Si")))


ensanut18_4analysis <- ensanut18_4analysis |>
  select(upm:ent, region, est_dis:f_20mas, age_cat, sexo, urbanicity, education, ses,
         insurance, l_indigena, work_pwk, usa_lentes, disc_visual, disc_visual_severa, 
         disc_visual_severa_num, cesd_score, cesd, suicide_int, suicide_int_py, suicide_int_py_r, 
         suicide_int_py_r_num, suicide_int_num, suicide_ideation, suicide_ideation_num)


save(ensanut18_4analysis, file = here("data/", "ensanut18_4analysis.RData"))
 
```
