==> code/3 - preliminary_analysis_abstract.Rmd <==
---
title: "3 - Preliminary analysis for abstract"
author: "Luis Segura"
date: "`r Sys.Date()`"
output: 
  html_document: 
    toc: true
    toc_float: true
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)

options(tinytex.verbose = TRUE, scipen = 999)

mypckgs <- c("tidyverse", "here", "survey", "srvyr")

for (package in mypckgs) {
  library(package, character.only = T)
}

load(here("data/", "ensanut18_4analysis.RData"))
```

```{r}
### creamos el objeto con el diseno muestral de la ensanut para calcular prevalencias ponderadas
diseno_ensanut <- ensanut18_4analysis |>
  as_survey_design(strata = est_dis, 
                   ids = upm_dis, 
                   weights = f_20mas, 
                   nest = T)
```

# Metodos
Usamos los datos de la encuesta nacional de salud y nutricion (ENSANUT) 2018 para calcular la prevalencia de discapacidad visual severa y sintomas depresivos en poblacion mexicana estratificando por grupos de edad y, ademas, por grupos de edad y subgrupos vulnerables. 

Para el calculo de las prevalencias incorporamos el diseno muestral y ponderadores de la ENSANUT, la cual es una muestra representativa de la poblacion mexicana. Estimamos la asociacion entre discapacidad visual severa y sintomas depresivos por medio de razones de prevalencias (RP) y sus intervalos de confianza (IC 95%). Para obtener las RP e IC 95%, ajustamos modelos de regresion tipo Poisson con errores estandar de tipo robusto e incorporamos el diseno muestral y ponderadores de la ENSANUT 2018. Ajustamos diferentes modelos estratificados por grupos de edad y, ademas, por grupos de edad y subgrupos vulnerables.

# Resultados
## Prevalencia de discapacidad visual y sintomas depresivos

En la poblacion mexicana la prevalencia de discapacidad visual es de `r diseno_ensanut |> group_by(disc_visual_severa) |> summarise(prop = survey_mean()) |> filter(disc_visual_severa == "Yes") |> mutate(prop = round(prop * 100, 2)) |> select(prop)`%. La prevalencia de sintomas depresivos es de `r diseno_ensanut |> group_by(cesd) |> summarise(prop = survey_mean()) |> filter(cesd == "Depressive symptoms") |> mutate(prop = round(prop * 100, 2)) |> select(prop)`%. La asociacion entre discapacidad visual severa y sintomas depresivos es de `r svyglm(cesd_num ~ disc_visual_severa, design = diseno_ensanut, family = poisson) |> emmeans::emmeans(trt.vs.ctrl ~ disc_visual_severa, data = ensanut18_4analysis, type = "response") |> pluck("contrasts") |> broom::tidy(conf.int = T) |> mutate(beta = str_glue("{round(ratio, 2)} (IC 95% = {round(conf.low, 2)}; {round(conf.high, 2)})")) |> select(beta)`.


## Grupos de edad{.tabset}
### Prevalencias
```{r}
my_fun <- function(x, y, z){
  
  my_formula <- as.formula(str_glue("{y} ~ {z} * {x}"))
  
  table <- diseno_ensanut |>
  group_by(!!sym(x), disc_visual_severa, cesd) |>
  summarise(prop = survey_mean(vartype = "ci")) |>
  ungroup() |>
  filter(cesd == "Depressive symptoms") |>
  mutate(prevalencia = str_glue("{round(prop * 100, 2)}%"), 
         `IC 95%` = str_glue("({round(prop_low * 100, 2)}; {round(prop_upp * 100, 2)})")) |>
  select(-c(cesd, prop, prop_low, prop_upp))
  
  fit <- svyglm(my_formula, design = diseno_ensanut, family = quasipoisson) 
  
  betas <- fit |>
  emmeans::emmeans(trt.vs.ctrl ~ disc_visual_severa, by = c(x), data = ensanut18_4analysis, type = "response") |>
  pluck("contrasts") |>
  broom::tidy(conf.int = T) |>
  mutate(`IC 95%` = str_glue("({round(conf.low, 2)}; {round(conf.high, 2)})"), 
         `valor p` = round(p.value, 4), 
         RP = round(ratio, 2))
  
  my_list <- list(table, betas)
  
  return(my_list)
  
}

results_1 <- my_fun("age_cat", "cesd_num", "disc_visual_severa")


results_1[[1]] |>
  rename(`grupo edad` = age_cat, 
         `discapacidad visual severa` = disc_visual_severa) |>
  DT::datatable(filter = "top")
```

### Razones de prevalencias
```{r}
results_1[[2]] |>
  select(`grupo_edad` = age_cat, `discapacidad visual severa` = contrast, RP, `IC 95%`, `valor p`) |>
  DT::datatable(filter = "top")
  
```

## Sexo{.tabset}
### Prevalencias
```{r}
results_2 <- my_fun("sexo", "cesd_num", "disc_visual_severa")

results_2[[1]] |>
  rename(`discapacidad visual severa` = disc_visual_severa) |>
  DT::datatable(filter = "top")
```


### Razones de prevalencia
```{r}
results_2[[2]] |>
  select(sexo, `discapacidad visual severa` = contrast, RP, `IC 95%`, `valor p`) |>
  DT::datatable(filter = "top")
```


## Estatus socioeconomico{.tabset}
### Prevalencias
```{r}
results_3 <- my_fun("ses", "cesd_num", "disc_visual_severa")

results_3[[1]] |>
  rename(`discapacidad visual severa` = disc_visual_severa, 
         `estatus socioeconomico` = ses) |>
  DT::datatable(filter = "top")
```


### Razones de prevalencia
```{r}
results_3[[2]] |>
  select(`estatus socioeconomico` = ses, `discapacidad visual severa` = contrast, RP, `IC 95%`, `valor p`) |>
  DT::datatable(filter = "top")
```

## Lengua indigena{.tabset}
### Prevalencias
```{r}
results_4 <- my_fun("l_indigena", "cesd_num", "disc_visual_severa")

results_4[[1]] |>
  rename(`discapacidad visual severa` = disc_visual_severa,
         `habla lengua indigena` = l_indigena) |>
  DT::datatable(filter = "top")
```

### Razones de prevalencia
```{r}
results_4[[2]] |>
  select(`habla lengua indigena` = l_indigena, `discapacidad visual severa` = contrast, RP, `IC 95%`, `valor p`) |>
  DT::datatable(filter = "top")
```

## Region{.tabset}
### Prevalencias
```{r}
results_5 <- my_fun("region", "cesd_num", "disc_visual_severa")

results_5[[1]] |>
  rename(`discapacidad visual severa` = disc_visual_severa) |>
  DT::datatable(filter = "top")
```

### Razones de prevalencia
```{r}
results_5[[2]] |>
  select(region, `discapacidad visual severa` = contrast, RP, `IC 95%`, `valor p`) |>
  DT::datatable(filter = "top")
```

## Aseguranza{.tabset}
### Prevalencias
```{r}
results_6 <- my_fun("insurance", "cesd_num", "disc_visual_severa")

results_6[[1]] |>
  rename(`discapacidad visual severa` = disc_visual_severa, 
         aseguranza = insurance) |>
  DT::datatable(filter = "top")
```

### Razones de prevalencia
```{r}
results_6[[2]] |>
  select(insurance, `discapacidad visual severa` = contrast, RP, `IC 95%`, `valor p`) |>
  DT::datatable(filter = "top")
```

## Empleo{.tabset}
### Prevalencias
```{r}
results_7 <- my_fun("work_pwk", "cesd_num", "disc_visual_severa")

results_7[[1]] |>
  rename(`discapacidad visual severa` = disc_visual_severa, 
         `trabaja` = work_pwk) |>
  DT::datatable(filter = "top")
```

### Razones de prevalencia
```{r}
results_7[[2]] |>
  select(`trabaja` = work_pwk, `discapacidad visual severa` = contrast, RP, `IC 95%`, `valor p`) |>
  DT::datatable(filter = "top")
```

## Educacion{.tabset}
### Prevalencias
```{r}
results_8 <- my_fun("education", "cesd_num", "disc_visual_severa") 

results_8[[1]] |>
  rename(`discapacidad visual severa` = disc_visual_severa, 
         `educacion` = education) |>
  DT::datatable(filter = "top")
```

### Razones de prevalencia
```{r}
results_8[[2]] |>
  select(`educacion` = education, `discapacidad visual severa` = contrast, RP, `IC 95%`, `valor p`) |>
  DT::datatable(filter = "top")
```
