==> code/4 - poster_plots.Rmd <==
---
title: "4 - Results for poster CONGISP"
author: "Luis Segura"
date: '"`r Sys.Date()`"'
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)

options(tinytex.verbose = TRUE, scipen = 999)

mypckgs <- c("tidyverse", "here", "survey", "srvyr", "doParallel", "foreach", "extrafont", "showtext", "patchwork")

for (package in mypckgs) {
  library(package, character.only = T)
  
}

## setup parallel processing
no_cores <- detectCores() - 1  
registerDoParallel(cores = no_cores)
options(cores = no_cores)
getDoParWorkers()                   
getDoParName()

loadfonts(device = "win")

font_add(family = "Tahoma", regular = "C:/Windows/Fonts/tahoma.ttf")
showtext_auto()

load(here("data/", "ensanut18_4analysis.RData"))
```

```{r}
ensanut18_4analysis <- ensanut18_4analysis |>
  mutate(cesd_num = ifelse(cesd == "Depressive symptoms", 1, 
                           ifelse(cesd == "No depressive symptoms", 0, NA)))

### creamos el objeto con el diseno muestral de la ensanut para calcular prevalencias ponderadas
diseno_ensanut <- ensanut18_4analysis |>
  as_survey_design(strata = est_dis, 
                   ids = upm_dis, 
                   weights = f_20mas, 
                   nest = T)
```

```{r}
my_fun <- function(x, y, z){
  
  my_formula <- as.formula(str_glue("{y} ~ {z} * {x}"))
  
  table <- diseno_ensanut |>
  group_by(!!sym(x), disc_visual_severa, cesd) |>
  summarise(prop = survey_mean(vartype = "ci")) |>
  ungroup() |>
  filter(cesd == "Depressive symptoms") |>
  mutate(prevalencia = str_glue("{round(prop * 100, 2)}%"), 
         `IC 95%` = str_glue("({round(prop_low * 100, 2)}; {round(prop_upp * 100, 2)})")) |>
  select(-c(cesd, prop, prop_low, prop_upp))
  
  fit <- svyglm(my_formula, design = diseno_ensanut, family = quasipoisson) 
  
  fit_emm <- fit |>
  emmeans::emmeans(trt.vs.ctrl ~ disc_visual_severa, by = c(x), data = ensanut18_4analysis, type = "response") 
  
  contrasts_emm <- fit_emm |>
  pluck("contrasts") |>
  broom::tidy(conf.int = T) |>
  mutate(var = x) |>
  select(var, level = !!sym(x), contrast, PR = ratio, conf.low, 
           conf.high, `p value` = ends_with("p.value"))
  
  my_list <- list(table = table, betas = contrasts_emm)
  
  return(my_list)
  
}

```

```{r}
my_vars <- names(ensanut18_4analysis)[c(13, 6, 16, 18, 8, 17, 19, 15)]

results_list <- foreach(i = my_vars, 
                        .packages = c("tidyverse", "survey", "srvyr", "emmeans")) %dopar% {
  
  my_fun(i, "cesd_num", "disc_visual_severa")
  
}

```


```{r}
overall_pr <- svyglm(cesd_num ~ disc_visual_severa, design = diseno_ensanut, family = poisson) |> 
  emmeans::emmeans(trt.vs.ctrl ~ disc_visual_severa, data = ensanut18_4analysis, type = "response") |> 
  pluck("contrasts") |> 
  broom::tidy(conf.int = T) |>
  mutate(var = "Overall", 
         level = "SVI (Yes/No)") |> 
  select(var, level, contrast, PR = ratio, conf.low, conf.high)
```


```{r}
betas_plot <- map(results_list, "betas") |>
  bind_rows() |>
    mutate(contrast = str_trim(contrast, side = "both"))

betas_plot <- overall_pr |>
  bind_rows(betas_plot)



plot_poster <- betas_plot |>
  mutate(term = case_when(level == "SVI (Yes/No)" ~ "Discapacidad visual \nsevera (Si/No) \n", 
                          level == "20 / 39 yo" ~ "20 a 39 años", 
                          level == "40 / 59 yo" ~ "40 a 59 años", 
                          level == "60+ yo" ~ "60+ años", 
                          level == "Men" ~ "Hombre", 
                          level == "Women" ~ "Mujer", 
                          level == "High" ~ "Alto", 
                          level == "Middle" ~ "Medio", 
                          level == "Low" ~ "Bajo", 
                          level == "No" ~ "No", 
                          level == "Yes" ~ "Si", 
                          level == "Center" ~ "Centro", 
                          level == "North" ~ "Norte", 
                          level == "Mexico City" ~ "Ciudad de México", 
                          level == "South" ~ "Sur", 
                          level == "No insurance" ~ "Sin servicio médico", 
                          level == "Insurance" ~ "Con servicio médico", 
                          level == "More than high school education" ~ "Mayor a preparatoria", 
                          level == "High school education" ~ "Preparatoria", 
                          level == "Secondary school" ~ "Secundaria", 
                          level == "Elementary education or lower" ~ "Primaria o menor"),
         term = factor(term), 
         orden = seq(1, length(var), 1),
         color = factor(c(1, rep(2, 3), rep(3, 2), rep(4, 3), rep(5, 2), 
                   rep(6, 4), rep(7, 2), rep(8, 2), rep(9, 4)), 
                   labels = c("Total", "Grupo de edad", "Sexo", "Nivel socioeconómico", "Lengua indigena", 
                              "Región", "Acceso a servicios médicos", "Empleado", "Educación"))) |>
  ggplot(aes(y = reorder(ordered(interaction(term, color, sep = "!")), desc(orden)), 
             x = PR, xmin = conf.low, xmax = conf.high, color = color)) +
  geom_point(size = 8, shape = "diamond") +
  geom_errorbarh(height = 0.3) + 
  scale_x_continuous(limits = c(1, 5.0), breaks = seq(1.0, 5.0, 1.0), name = "Razones de Prevalencia (IC 95%)")  +
  scale_y_discrete(guide = ggh4x::guide_axis_nested(delim = "!")) +
  geom_vline(xintercept = 1, color = "black", linetype = "solid", alpha = 0.5) +
  ggthemes::theme_calc() +
  theme(legend.position = "none", 
        text = element_text(size = 220, family = "Tahoma"), 
        # plot.margin = margin(20, 20, 20, 20), 
        panel.background = element_rect(fill = "transparent", color = NA), 
        plot.background = element_rect(fill = "transparent", color = NA), 
        axis.text.y = element_text(margin = margin(l = 4, r = 4, unit = "pt")), 
        ggh4x.axis.nesttext.y = element_text(size = 220, face = "bold")) +
  labs(y = "",
       title = "") 

# figure_1 <- plot_poster + plot_annotation(title = "Figura 1. Razones de prevalencia de síntomas depresivos comparando aquellos con y sin discapacidad visual severa en la población total \ny por grupos sociodemográficos en adultos Mexicanos. ENSANUT 2018.", 
#                                    theme = theme(plot.title = element_text(hjust = 0, size = 28, face = "bold", family = "Tahoma"),
#       plot.margin = margin(5, 20, 20, 5)))


ggsave(here("results/", "forestplot_poster.jpeg"), plot = plot_poster, 
       width = 12, height = 10, units = "in", dpi = 1800)

```
